+++ apps/v2/components/client.ts
@@ -8,3 +8,5 @@ export {
   NavMenu,
   type MenuItemProps,
 } from '@maker-ui/layout/dist/client'
+
+export { Modal } from '@maker-ui/modal'