+++ packages/spinners/tsup.config.ts
@@ -1,0 +1,14 @@
+import { defineConfig } from 'tsup'
+import { buildConfig } from '@maker-ui/build-tools'
+
+const config = buildConfig({
+  clean: false,
+  minifyIdentifiers: true,
+  minifyWhitespace: true,
+  minifySyntax: true,
+  entry: {
+    index: 'src/index.ts',
+  },
+})
+
+export default defineConfig(config)